### 2019-09-29 #1772 @danielskatz
@ramiyaari - I see some high-level issues here:

* no readme
* no license

Please see https://joss.readthedocs.io/en/latest/submitting.html#typical-paper-submission-flow

Also, in your paper, you say the bib is "bibliography: simode.bib" but it's actually "paper.bib" - please fix this, then enter ```@whedon generate pdf``` as a new comment in this issue to recompile.
