### 2019-09-30 2002640
removed unsupported latex math command
