### 2019-09-30 43484e9
removed more unsupported latex math commands
