### 2019-09-30 aa4806f
separated eq 6 and 7
