### 2019-09-30 #1772 @ramiyaari
@danielskatz thanks for your comments
I believe I fixed all these issues
