### 2019-09-30 #1772 @ramiyaari
Please update the software version to 1.1.6
Suggested Reviewers: Jgoldfar, trashbirdecology, MikeLydeamore, MaaniBeigy, llauriciu, matloff, ellessenne, osorensen, RomainAzais
