### 2019-09-30 06744b1
fixed bug introduced in v1.1.5
