### 2019-10-01 #1772 @ramiyaari
There is a README.txt and a LICENSE.txt file under inst/joss in the repo, together with paper.md
