### 2019-10-03 #1772 @ramiyaari
@danielskatz: There is a README.txt and a LICENSE.txt file under inst/joss in the repo, together with paper.md
