### 2019-10-09 #1772 @labarba
👋 @mjsottile — Could you handle this submission as editor?
**simode: R Package for Statistical Inference of Ordinary Differential Equations using Separable Integral-Matching**
Repository: https://github.com/ramiyaari/simode/
