### 2019-10-11 #1772 @mjsottile
Yes, I can take this submission as editor.
