### 2019-10-16 #1772 @mjsottile
Other possible reviewers from the spreadsheet: @jgoldfar - you appear to have knowledge of differential equations and R. @fabian-s - you appear to have knowledge of R and inference methods.  Interested in taking this on to help review?
