### 2019-10-18 #1772 @jgoldfar
I would be able to take this on at the beginning of November; I'd be happy to help if there is still a need for review.
> On Oct 16, 2019, at 12:09 AM, Matthew Sottile <notifications@github.com> wrote:
>
> Other possible reviewers from the spreadsheet: @jgoldfar <https://github.com/jgoldfar> - you appear to have knowledge of differential equations and R. @fabian-s <https://github.com/fabian-s> - you appear to have knowledge of R and inference methods. Interested in taking this on to help review?
>
