### 2019-10-30 #1772 @mjsottile
Thanks @jgoldfar.  That timeline should be fine in November.
