### 2019-10-30 #1772 @MaaniBeigy
Yes, I am interested

On Wed, 30 Oct 2019, 07:33 Øystein Sørensen, <notifications@github.com>
wrote:

> Yes, I am interested.
>
