### 2019-10-30 #1850 @whedon
[REVIEW]: simode: R Package for Statistical Inference of Ordinary Differential Equations using Separable Integral-Matching
**Submitting author:** @ramiyaari (<a href="http://orcid.org/0000-0002-8808-8937">Rami Yaari</a>)
**Repository:** <a href="https://github.com/ramiyaari/simode/" target ="_blank">https://github.com/ramiyaari/simode/</a>
**Version:** v1.1.8
