### 2019-11-04 ramiyaari/simode#1 @osorensen
Review of paper submitted to JOSS
Attached is the pdf of the paper [submitted to JOSS](https://github.com/openjournals/joss-reviews/issues/1850) with my comments and remarks. You might need to open it with Acrobat Reader for the comments to show properly.

[10.21105.joss.01850_with_comments.pdf](https://github.com/ramiyaari/simode/files/3803251/10.21105.joss.01850_with_comments.pdf)
