### 2019-11-04 c7cd295
Corrected small typo in documentation
