### 2019-11-04 ramiyaari/simode#3 @osorensen
Automated tests
One of the points on my [review checklist](https://github.com/openjournals/joss-reviews/issues/1850) asks:

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

I am not able to find any such tests in this package. I suggest adding unit tests using the `devtools` package. A good place to start is reading [this chapter](http://r-pkgs.had.co.nz/tests.html).
