### 2019-11-04 ramiyaari/simode#4 @osorensen
Community guidelines
Considering the following point on the [review checklist](https://github.com/openjournals/joss-reviews/issues/1850) for JOSS:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please add a note about how to do this, either as part of the README file, or as a separate file in the repository.
