### 2019-11-09 ramiyaari/simode#5 @cscherrer
Example too wide for page
Hello,

You example includes the following lines:
```R
R> est_lin <- simode(
+ equations=equations, pars=lin_pars, fixed=c(x0,theta[nlin_pars]), time=time,
R> summary(est_lin)
```
The second seems too wide for the page, and needs a line break.
