### 2019-11-16 9a041dd
Update paper.md
fixes to comments by @oyss
