+++ inst/joss/paper.bib
@@ -98,10 +98,9 @@ year={2018}
 
 @book{anderson1992infectious,
   title={Infectious diseases of humans: dynamics and control},
-  author={Anderson, Roy M and May, Robert M and Anderson, B},
-  volume={28},
+  author={Anderson, Roy M and May, Robert M},
   year={1992},
-  publisher={Wiley Online Library}
+  publisher={Oxford University Press}
 }
 
 
@@ -115,7 +114,7 @@ year={2018}
 
 @article{hooker2015collocinfer,
   title={CollocInfer: Collocation Inference in Differential Equation Models},
-  author={Hooker, Giles and Ramsay, James O and Xiao, Luo and others},
+  author={Hooker, Giles and Ramsay, James O and Xiao, Luo},
   journal={Journal of Statistical Software},
   year={2015}
 }
