+++ inst/joss/paper.bib
@@ -1,6 +1,6 @@
 
 @Article{deSolve,
-    title = {Solving Differential Equations in R: Package deSolve},
+    title = {Solving Differential Equations in {R}: Package deSolve},
     author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer},
     journal = {Journal of Statistical Software},
     volume = {33},
