+++ inst/joss/paper.bib
@@ -24,7 +24,7 @@
 }
 
 @Manual {simode,
-title={simode: R Package for Statistical Inference of Ordinary Differential Equations using Separable Integral-Matching},
+title={{simode}: R Package for Statistical Inference of Ordinary Differential Equations using Separable Integral-Matching},
 author={Yaari, Rami  and Dattner, Itai },
 year={2018},
 url={https://cran.r-project.org/web/packages/simode/vignettes/R_package_simode.pdf}
