### 2019-11-16 2345a4f
Update paper.md
