### 2019-11-16 ramiyaari/simode#1 @ramiyaari
Thanks @osorensen , I believe I handled all your comments in this new version.
