### 2019-11-16 ramiyaari/simode#3 @ramiyaari
There are demos as part of the package which you can run.
To see the avaliable demos type: demo(package='simode')
To run a demo type: demo(<demo_name>,package='simode')
Is this not good enough for automated tests?
