### 2019-11-16 ramiyaari/simode#3 @osorensen
The demos do a great job in demonstrating how to use the package, but will they fail if a subtle bug causes one of the package functions to return the wrong answer, but without issuing an error or warning?

When reading the point on the review checklist, I had in mind something like the following example. I have taken most of the code from the example to the function `simode()`, but added some lines at the bottom showing how one could test the numerical output from the function. As far as I can see, the demos do not have any verification that the numbers coming out are correct.

```r
library(simode)

######## THE FOLLOWING LINES ARE TAKEN FROM THE EXAMPLE TO SIMODE() ########
## generate model equations and parameters (X=Prey,Y=Predator)
pars <- c('alpha','beta','gamma','delta')
vars <- c('X','Y')
eq_X <- 'alpha*X-beta*X*Y'
eq_Y <- 'delta*X*Y-gamma*Y'
equations <- c(eq_X,eq_Y)
names(equations) <- vars
x0 <- c(0.9,0.9)
names(x0) <- vars
theta <- c(2/3,4/3,1,1)
names(theta) <- pars

## generate observations
n <- 50
time <- seq(0,25,length.out=n)
model_out <- solve_ode(equations,theta,x0,time)
x_det <- model_out[,vars]
set.seed(1000)
sigma <- 0.05
obs <- list()
for(i in 1:length(vars)) {
  obs[[i]] <- pmax(0, rnorm(n,x_det[,i],sigma))
}
names(obs) <- vars

## estimate model parameters with known initial conditions
simode_fit1 <- simode(equations=equations, pars=pars, fixed=x0, time=time, obs=obs)

######## THE FOLLOWING LINES SHOW HOW ONE COULD PERFORM UNIT TESTING OF THE SIMODE() FUNCTION ########
testthat::test_that("simode returns correct answers",{
  testthat::expect_equal(round(simode_fit1$nls_loss, 5), 0.24645)
  testthat::expect_equal(as.numeric(round(simode_fit1$nls_sol[50, ], 5)), c(0.56900, 0.63472 ))
}
)
```

Whether this type of testing is necessary or not depends on how one interprets the point on the reviewer checkelist. Perhaps editor @mjsottile could clarify if demos are sufficient?
