### 2019-11-17 ramiyaari/simode#4 @ramiyaari
Hi @osorensen , regarding this point, we are not sure exactly what should be added to the README as we are not aware of the standard practices on github. We thought that the fact the project is on github and with the license that we've supplied means that anyone could use it and download the code and make changes. It would be great if anybody would want to build on this project and contribute to it. What is the standard way to reflect that? Do you have any example from a different project?
