### 2019-11-17 ramiyaari/simode#4 @osorensen
I agree with you that being on GitHub should be sufficient. This is fine with me.
