### 2019-11-17 ramiyaari/simode#4 @osorensen
@ramiyaari , in my own package `hdme`, which was published in JOSS, I added the following section to the GitHub readme. Feel free to use it: https://github.com/osorensen/hdme#contributions
