### 2019-11-18 ramiyaari/simode#4 @ramiyaari
@osorensen, thankyou very much. I've updated the simode readme accordingly.
