### 2019-11-18 ramiyaari/simode#3 @mjsottile
Based on the guidance in the JOSS documentation, the testing aspect reads as follows:

Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

- **Good**: An automated test suite hooked up to an external service such as Travis-CI or similar
- **OK**: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
- **Bad (not acceptable)**: No way for you the reviewer to objectively assess whether the software works

I would interpret the demos included in the package as falling in between the second bullet and the third - "OK" and "Bad".  On the OK side, they do serve the role of demonstrating how the code works, and represent as basic set of specific test cases that can be run manually.  On the Bad side, my browsing of the code doesn't show that the user has any way to assess what a correct output is, so they cannot assess whether or not the code executed correctly or not.  They can only observe the code executing and producing some output - correct or not.

I agree with @osorensen - the devtools package does provide some basic mechanisms for automatically running tests, including automatic assessment of whether they should be considered passing or not passing.  I would recommend adapting the demos to be usable through that mechanism.  I would also recommend improving the documentation in the demos if they are intended to be usable as tests to clearly articulate what the expected result of running the code should be.
