### 2019-11-20 ramiyaari/simode#3 @ramiyaari
OK, I will work on adding automated tests soon
