### 2019-11-26 dd0a95f
Added unit tests using testthat
