### 2019-11-26 ramiyaari/simode#3 @ramiyaari
OK @osorensen  and @mjsottile , I've added unit test using testthat to the package
