### 2019-11-28 ramiyaari/simode#3 @osorensen
Thanks, I think the tests look very satisfactory. I do however get a warning when testing the package, caused by the following lines in `test_simode_argument.R`

```
obs <- list()
obs['X'] <- 1:10
obs['Y'] <- 10:1
```

If the assignments are changed to the following, everything will work fine, so please do that.

```
obs <- list()
obs['X'] <- 1:10
obs['Y'] <- 10:1
```
