### 2019-11-30 e9fd919
Fixed problem with using decouple_equations=T with fixed initial condition Updated test files Modified License and Readme files
