+++ tests/testthat/test_simode_decouple.R
@@ -39,14 +39,14 @@ set.seed(1000)
 sigma <- 0.01
 obs.list <- c()
 N <- 3
-x0_vals <- matrix(NA,N,M)
-colnames(x0_vals) <- vars
+x0.list <- list()
 for(j in 1:N) {
   x0 <- round(rnorm(M,0.5,0.3),2)
   x0[x0<0] <- 0
   x0[x0>1] <- 1
-  x0_vals[j,] <- x0
-  model_out <- solve_ode(equations,theta,x0_vals[j,],time)
+  names(x0) <- vars
+  x0.list[[j]] <- x0
+  model_out <- solve_ode(equations,theta,x0.list[[j]],time)
   x_det <- model_out[,vars]
   obs <- list()
   for(i in 1:length(vars)) {
@@ -64,35 +64,59 @@ names(pars_max) <- vars
 b_init <- rep(1,M)
 names(b_init) <- b_names
 
-fits <- simode(
-          equations=equations, pars=c(vars,pars), nlin_pars=b_names, start=b_init,
+fits1 <- simode(
+          equations=equations, pars=pars, nlin_pars=b_names, start=b_init, fixed=x0.list,
           time=time, obs=obs.list, obs_sets=N, lower=pars_min, upper=pars_max, decouple_equations = T,
           simode_ctrl = simode.control(optim_type='im',obs_sets_fit='separate_x0'))
 
+fits2 <- simode(
+  equations=equations, pars=c(pars,vars), nlin_pars=b_names, start=b_init,
+  time=time, obs=obs.list, obs_sets=N, lower=pars_min, upper=pars_max, decouple_equations = T,
+  simode_ctrl = simode.control(optim_type='im',obs_sets_fit='separate_x0'))
+
 
 test_that("simode multi call with decoupling returns expected format and values", {
-  expect_is(fits,'list.simode')
-  expect_equal(length(fits),N)
+  expect_is(fits1,'list.simode')
+  expect_equal(length(fits1),N)
   for(i in 1:N) {
-    fit <- fits[[i]]
+    fit <- fits1[[i]]
     obs <- obs.list[[i]]
-    x0 <- x0_vals[i,]
     expect_is(fit,'simode')
     expect_equal(fit$equations,equations)
-    expect_equal(fit$pars,c(vars,pars))
+    expect_equal(fit$pars,pars)
     expect_equal(fit$time,time)
     expect_equal(fit$obs,obs)
     expect_true(is.numeric(fit$im_loss))
     expect_true(is.numeric(fit$im_pars_est))
-    expect_equal(names(fit$im_pars_est),c(vars,pars))
+    expect_equal(names(fit$im_pars_est),pars)
   }
-  x0_est <- summary(fits)$im_est[,vars]
+  a_est <- colMeans(summary(fits1)$im_est[,a_names])
+  expect_true(mean(abs((a_est-theta[a_names])))<0.05)
+  b_est <- colMeans(summary(fits1)$im_est[,b_names])
+  expect_true(mean(abs((b_est-theta[b_names])))<0.2)
+
+  expect_is(fits2,'list.simode')
+  expect_equal(length(fits2),N)
+  for(i in 1:N) {
+    fit <- fits2[[i]]
+    obs <- obs.list[[i]]
+    expect_is(fit,'simode')
+    expect_equal(fit$equations,equations)
+    expect_equal(fit$pars,c(pars,vars))
+    expect_equal(fit$time,time)
+    expect_equal(fit$obs,obs)
+    expect_true(is.numeric(fit$im_loss))
+    expect_true(is.numeric(fit$im_pars_est))
[…]
