### 2019-11-30 ramiyaari/simode#3 @ramiyaari
Thanks, I've fixed the test file (plus fixed a problem I found on the way in the code).
