### 2019-12-02 ramiyaari/simode#3 @mjsottile
@osorensen are you satisfied that the test issue is addressed for the review to proceed forward?
