+++ README.md
@@ -64,6 +64,6 @@ For support and bug reports send an email to: ramiyaari@gmail.com or open an iss
 Code contributions to simode are also very welcome.
 
 #### References:
-Dattner Itai, Klaassen Chris. <b>Optimal rate of direct estimators in systems of ordinary differential equations linear in functions of the parameters. </b><i>Electonic Journal of Statistics</i>, 9(2), 2015. [<a href="doi:10.1214/15-EJS1053">link</a>]
+Dattner Itai, Klaassen Chris. <b>Optimal rate of direct estimators in systems of ordinary differential equations linear in functions of the parameters. </b><i>Electonic Journal of Statistics</i>, 9(2), 2015. [<a href="https://projecteuclid.org/euclid.ejs/1440680332">link</a>]
 
-Dattner Itai, Miller Ezer, Petrenko Margarita, Kadouri Daniel, Jurkevitch Edouard and Huppert Amit. <b>Modelling and parameter inference of predator–prey dynamics in heterogeneous environments using the direct integral approach. </b><i>Journal of the Royal Society Interface</i>, 14(126), 2017. [<a href="doi:10.1098/rsif.2016.0525">link</a>].
+Dattner Itai, Miller Ezer, Petrenko Margarita, Kadouri Daniel, Jurkevitch Edouard and Huppert Amit. <b>Modelling and parameter inference of predator–prey dynamics in heterogeneous environments using the direct integral approach. </b><i>Journal of the Royal Society Interface</i>, 14(126), 2017. [<a href="https://royalsocietypublishing.org/doi/10.1098/rsif.2016.0525">link</a>].
