### 2019-12-06 #1850 @mjsottile
Can the reviewers @cscherrer, @jgoldfar, @osorensen confirm their recommendation to accept the submission by responding in this thread?

Also, @ramiyaari - I see that there is a pull request still waiting for a documentation correction.  Can you address that and close the open pull request?
