### 2019-12-08 #1850 @ramiyaari
Thanks @mjsottile . I've merged the pull request.
