### 2019-12-09 #1850 @jgoldfar
@mjsottile I confirm my recommendation
> On Dec 6, 2019, at 5:14 PM, Matthew Sottile <notifications@github.com> wrote:
>
> Can the reviewers @cscherrer <https://github.com/cscherrer>, @jgoldfar <https://github.com/jgoldfar>, @osorensen <https://github.com/osorensen> confirm their recommendation to accept the submission by responding in this thread?
>
> Also, @ramiyaari <https://github.com/ramiyaari> - I see that there is a pull request still waiting for a documentation correction. Can you address that and close the open pull request?
>
