### 2019-12-13 a4902e7
minor editorial fixes for final version of paper
