### 2019-12-13 #1850 @mjsottile
Only a few steps left to finalize acceptance @ramiyaari

1. Please review the pull request I just submitted.  I did a final pass over the paper and found a small number of minor changes to the text.
2. Please add the missing DOIs to the references that lack them. See the message from whedon above.
3. Once you have completed this, the last step is to make a tagged release and archive.  Please comment in this thread with the version number and archive DOI.

Once those three steps are completed, I will send the submission off to the JOSS EICS for final processing.
