### 2019-12-14 7a44233
Update paper.bib
added missing doi
