+++ inst/joss/paper.bib
@@ -37,7 +37,8 @@ url={https://cran.r-project.org/web/packages/simode/vignettes/R_package_simode.p
   journal={ISRN Biomathematics},
   volume={2013},
   year={2013},
-  publisher={Hindawi Publishing Corporation}
+  publisher={Hindawi Publishing Corporation},
+  doi = {10.1155/2013/897658}
 }
 
 @article{voit2015150,
@@ -48,7 +49,8 @@ url={https://cran.r-project.org/web/packages/simode/vignettes/R_package_simode.p
   number={1},
   pages={e1004012},
   year={2015},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi = {10.1371/journal.pcbi.1004012}
 }
 
 @article{dattner2017modelling,
@@ -57,9 +59,9 @@ url={https://cran.r-project.org/web/packages/simode/vignettes/R_package_simode.p
   journal={Journal of The Royal Society Interface},
   volume={14},
   number={126},
-  pages={20160525},
   year={2017},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi = {10.1098/rsif.2016.0525}
 }
 
 @article{vujavcic2018consistency,
@@ -69,7 +71,8 @@ url={https://cran.r-project.org/web/packages/simode/vignettes/R_package_simode.p
   volume={132},
   pages={40--45},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.spl.2017.08.013}
 }
 
 @article{yaarietal18,
@@ -93,7 +96,8 @@ journal = {Statistica Neerlandica},
 volume = {72},
 number = {2},
 pages = {126-156},
-year={2018}
+year={2018},
+doi = {10.1111/stan.12124}
 }
 
 @book{anderson1992infectious,
@@ -116,7 +120,8 @@ year={2018}
   title={CollocInfer: Collocation Inference in Differential Equation Models},
   author={Hooker, Giles and Ramsay, James O and Xiao, Luo},
   journal={Journal of Statistical Software},
-  year={2015}
+  year={2015},
+  doi = {10.18637/jss.v075.i02}
 }
 
 @article{vujavcic2015time,
@@ -127,7 +132,8 @@ year={2018}
   number={6},
   pages={1057--1070},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s11222-014-9486-9}
 }
 
 @article{blind,
@@ -163,7 +169,8 @@ year={2018}
   number={4},
   pages={1176--1184},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/biom.12348}
 }
 
 @article{hooker2011parameterizing,
@@ -174,7 +181,8 @@ year={2018}
   number={60},
   pages={961--974},
   year={2011},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi = {10.1098/rsif.2010.0412}
 }
 
 
@@ -217,7 +225,8 @@ year={2018}
   number={1},
   pages={112--125},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.jmva.2008.03.005}
 }
 
 @article{hooker2009forcing,
@@ -228,7 +237,8 @@ year={2018}
   number={3},
   pages={928--936},
   year={2009},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.1541-0420.2008.01172.x}
 }
 
 
@@ -240,7 +250,8 @@ year={2018}
   number={11},
[…]
