### 2019-12-14 #1850 @ramiyaari
Hi @mjsottile , I've completed the first two tasks in your list. Regarding the third, if I understand what you've asked for then the current version of the software in CRAN is 1.1.8, and the doi of the joss paper is 10.21105/joss.01850.
Thanks
