### 2019-12-14 #1850 @danielskatz
For the second part, JOSS needs the software to be archived in an archival repository, such as Zenodo, figshare, or an institutional repository.  Once this is done, that repository should provide a DOI to the archive, which is what @mjsottile needs.  For software in GitHub, the easiest path is usually to use to the [Zenodo-GitHub linkage](https://guides.github.com/activities/citable-code/), being sure to manually change the metadata in Zenodo so that the title and authors of the archive match the those of the JOSS paper.
