### 2019-12-19 #1850 @mjsottile
:wave: @openjournals/joss-eics This paper has completed the review process and I believe it's ready for final processing.  Let me know if there is anything additional that I should do.
