### 2019-12-19 #1850 @labarba
@ramiyaari — I'm the associate EiC on rotation, and will process your paper to publication. We ask authors to edit the metadata (in this case, the title) of the Zenodo deposit so it matches the JOSS paper. Could you make that change?
