### 2019-12-19 #1850 @labarba
Looking at the paper, it does look rather long for the typical JOSS paper. In fact, the [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html) stipulates that papers should be _short_: "The paper should be between 250-1000 words." And also,

> a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

Do you think the detailed example should rather be in the documentation, and simply be mentioned here with a link?
