+++ inst/joss/paper.md
@@ -1,0 +1,207 @@
+---
+title: 'simode: R Package for Statistical Inference of Ordinary Differential Equations using Separable Integral-Matching'
+authors:
+- affiliation: 1, 2
+  name: Rami Yaari
+- affiliation: 1
+  name: Itai Dattner
+date: "23 September 2019"
+output:
+  html_document:
+    df_print: paged
+  pdf_document: default
+bibliography: paper.bib
+tags:
+- R
+- dynamic systems
+- ordinary differential equations
+- statistical inference
+- integral-matching
+- separable least-squares
+affiliations:
+- index: 1
+  name: Department of Statistics, University of Haifa, Haifa, 34988, Israel
+- index: 2
+  name: Bio-statistical and Bio-mathematical Unit, The Gertner Institute for Epidemiology
+    and Health Policy Research, Chaim Sheba Medical Center, Tel Hashomer, 52621, Israel
+---
+
+# Summary
+
+Systems of ordinary differential equations (ODEs) are commonly used for mathematical modeling of the rate of change of dynamic processes in areas such as mathematical biology [@edelstein2005mathematical], biochemistry [@voit2000computational] and compartmental models in epidemiology [@anderson1992infectious], to mention a few. Inference of ODEs involves the 'standard' statistical problems such as studying the identifiability of a model, estimating model parameters, predicting future states of the system, testing hypotheses, and choosing the 'best' model. However, dynamical systems are typically very complex: nonlinear, high dimensional and only partialy measured. Moreover, data may be sparse and noisy. Thus, statistical learning (inference, prediction) of dynamical systems is not a trivial task in practice. In particular, numerical application of standard estimators, like maximum-likelihood or least-squares, may be difficult or computationally costly. It typically requires solving the system numerically for a large set of potential parameters values, and choosing the optimal values using some nonlinear optimization technique. Starting from a random initial guess, the optimization can take a long time to converge to the optimal solution. Furthermore, there is no guarantee the optimization will converge to the optimal solution at all.   
+
+``simode`` is an R package for conducting statistical inference for ordinary differential equations that aims to ease the optimization process and provide more robust solutions to parameter estimation problems. The package implements a 'two-stage' approach. In the first stage, fast estimates of the ODEs' parameters are calculated by way of minimization of an integral criterion function while taking into account separability of parameters and equations (if such a mathematical feature exists). In the second stage, a regular nonlinear least-squares optimization is performed starting from the estimates obtained in the first stage, in order to try and improve these estimates.
+
[…]
