### 2019-12-22 #1850 @ramiyaari
Hi @labarba, I've fixed the title of the Zenodo to match that of the JOSS paper.
I've also edited the paper and reduced the space taken by the example by two pages.
If that is still too long than I will remove the example completely,
but I was looking for example at this paper (https://joss.theoj.org/papers/10.21105/joss.01601)
and saw it is similar in length to what we have now.
