### 2019-12-22 #1850 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
